Read an MDN with mixed case content type, like "TeXt/PlaIn"
http://www.w3.org/Protocols/rfc1341/4_Content-Type.html

The type, subtype, and parameter names are not case sensitive. For
example, TEXT, Text, and TeXt are all equivalent. Parameter values are
normally case sensitive, but certain parameters are interpreted to be
case- insensitive, depending on the intended use. (For example,
multipart boundaries are case-sensitive, but the "access- type" for
message/External-body is not case-sensitive.)